Rename to do_and_temp_meshgrid
